fix database scheme

+++ src/drizzle/schema/single-player-game.ts
@@ -1,4 +1,4 @@
-import { pgTable, uuid, numeric } from "drizzle-orm/pg-core";
+import { pgTable, uuid, integer } from "drizzle-orm/pg-core";
 import { createdAt, id } from "../schema-helpers";
 import { UsersTable } from "./users";
 import { relations } from "drizzle-orm";
@@ -6,8 +6,8 @@ import { relations } from "drizzle-orm";
 export const SinglePlayerGamesTable = pgTable("singleplayer_games", {
     id,
     userHostId: uuid().references(() => UsersTable.id),
-    currentRound: numeric(),
-    totalRounds: numeric(),
+    currentRound: integer(),
+    totalRounds: integer(),
     createdAt
 });
